==> FW/Scripts/1. FW_airports_merging.ipynb <==
import os
import sys
import numpy as np
import pandas as pd
from geopy import Point
from geopy import distance
pd.set_option('display.max_columns', 999)

# import datasets
fw = pd.read_csv("..//Datasets/NEW/FW_flights.csv", parse_dates=['date'])
fw_info = pd.read_csv("..//Datasets/NEW/FW_services_info.csv")
airports = pd.read_csv("..//Datasets/Airports/BTS_Airports_LAT_LON.csv")
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# number of existing tail numbers with flight history
fw.tail_number.nunique() 

# The percentage of the total tail numbers
print(round(fw.tail_number.nunique() / fw_info.Tail_number.nunique() * 100, 2), "%")

fw.info()

# check for null values
fw.isna().sum()

#### Data preprocessing
---

# check duplicates airport_codes
airport_codes.duplicated(subset=['gps_code']).sum()

# check duplicates airport_codes
airport_codes.duplicated(subset=['local_code']).sum()

# Remove useless words
fw = fw.replace(regex={'Near ':'', 'First seen ':'', 'Last seen ':'', 'result unknown': np.nan, 'En Route':None, 'unknown': np.nan,
                       'n/a': np.nan, 'Unknown':None, 'Diverted':np.nan, '¬†':' ', 'Scheduled': None, 'Cancelled': np.nan})

# count of training flights 
len(fw.loc[fw.origin == fw.destination])

# drop all training flights 
fw = fw.drop(fw.loc[fw.origin == fw.destination].index, axis=0)

# count of incomplete flights 
len(fw.loc[(fw.arrival.isna()) | (fw.duration.isna())])

# drop all incomplete flights 
fw = fw.drop(fw.loc[(fw.arrival.isna()) | (fw.duration.isna())].index, axis=0)

# split origin_Latitude and origin_Longitude
fw['origin_Latitude'] = fw[fw.origin_location.str.startswith('L ')].origin_location.replace(regex={'L ':''}).str.split(" ", n = 1, expand = True)[0]
fw['origin_Longitude'] = fw[fw.origin_location.str.startswith('L ')].origin_location.replace(regex={'L ':''}).str.split(" ", n = 1, expand = True)[1]

# split destination_Latitude and destination_Longitude
fw['destination_Latitude'] = fw[fw.destination_location.str.startswith('L ', na=False)].destination_location.replace(regex={'L ':''}).str.split(" ", n = 1, expand = True)[0]
fw['destination_Longitude'] = fw[fw.destination_location.str.startswith('L ', na=False)].destination_location.replace(regex={'L ':''}).str.split(" ", n = 1, expand = True)[1]

# convert to numeric
fw['origin_Latitude'] = pd.to_numeric(fw.origin_Latitude, errors='coerce')
fw['origin_Longitude'] = pd.to_numeric(fw.origin_Longitude, errors='coerce')

fw['destination_Latitude'] = pd.to_numeric(fw.destination_Latitude, errors='coerce')
fw['destination_Longitude'] = pd.to_numeric(fw.destination_Longitude, errors='coerce')

# split ICAO_code
fw['ICAO_code_origin'] = fw[fw.origin_Latitude.isna()].origin_location.str.rsplit(" - ", n = 1, expand = True)[1]
fw['ICAO_code_destination'] = fw[fw.destination_Latitude.isna()].destination_location.str.rsplit(" - ", n = 1, expand = True)[1]

# replace dual ICAO code
fw.loc[fw.ICAO_code_origin.str.contains(' / ', na=False), 'ICAO_code_origin'] = fw.loc[fw.ICAO_code_origin.str.contains(' / ', na=False), 'ICAO_code_origin'].str.split(" / ", expand=True, n=1)[0]

fw.loc[fw.ICAO_code_destination.str.contains(' / ', na=False), 'ICAO_code_destination'] = fw.loc[fw.ICAO_code_destination.str.contains(' / ', na=False), 'ICAO_code_destination'].str.split(" / ", expand=True, n=1)[0]

#### Origin Location
---

# merge origin airport information 1 time
fw = fw.merge(airport_codes[['ident']].add_suffix('_origin_1'),
         how='left', left_on='ICAO_code_origin', right_on='ident_origin_1')

# Switch dual ICAO codes
fw.loc[(fw.ICAO_code_origin.notnull()) & (fw.ident_origin_1.isna()) & fw.origin.str.contains(' / '), 'ICAO_code_origin'] = fw.loc[(fw.ICAO_code_origin.notnull()) & (fw.ident_origin_1.isna()) & fw.origin.str.contains(' / '), 'origin'].str.split(' / ', n=1, expand=True)[1].str.replace(')', '')

# merge origin airport information 2 time
fw = fw.merge(airport_codes[['ident']].add_suffix('_origin_2'),
         how='left', left_on='ICAO_code_origin', right_on='ident_origin_2')

# merge similar columns
fw.loc[(fw.ICAO_code_origin.notnull()) & (fw.ident_origin_1.isna()) & (fw.origin.str.contains(' / ')), 'ident_origin_1'] = fw.loc[(fw.ICAO_code_origin.notnull()) & (fw.ident_origin_1.isna()) & (fw.origin.str.contains(' / ')), 'ident_origin_2']

# Drop duplicated column
fw = fw.drop(['ident_origin_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# drop duplicates in gps_code
airport_codes = airport_codes.drop_duplicates(subset=['gps_code'])

# merge origin airport information 3 time
fw = fw.merge(airport_codes[['gps_code', 'ident']].add_suffix('_origin_2'),
         how='left', left_on='ICAO_code_origin', right_on='gps_code_origin_2')

# merge similar columns
fw.loc[(fw.ICAO_code_origin.notnull()) & (fw.ident_origin_1.isna()), 'ident_origin_1'] = fw.loc[(fw.ICAO_code_origin.notnull()) & (fw.ident_origin_1.isna()), 'ident_origin_2']

# Drop duplicated columns
fw = fw.drop(['ident_origin_2', 'gps_code_origin_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# drop duplicates in local_code
airport_codes = airport_codes.drop_duplicates(subset=['local_code'])

# merge origin airport information 4 times
fw = fw.merge(airport_codes[['local_code', 'ident']].add_suffix('_origin_2'),
         how='left', left_on='ICAO_code_origin', right_on='local_code_origin_2')

# merge similar columns
fw.loc[(fw.ICAO_code_origin.notnull()) & (fw.ident_origin_1.isna()), 'ident_origin_1'] = fw.loc[(fw.ICAO_code_origin.notnull()) & (fw.ident_origin_1.isna()), 'ident_origin_2']

# Drop duplicated columns
fw = fw.drop(['ident_origin_2', 'local_code_origin_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# split coordinates and convert to numeric
airport_codes["Latitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[1])
airport_codes["Longitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[0])

airport_codes = airport_codes.replace(regex={r' Heliport': '', r' Airpark':'', r' Airport': '', r' Regional':'', r' Municipal':'', r' International':'',
                                            r' LRRS':''})

airport_codes.loc[:, 'name'] = airport_codes.loc[:, 'name'] + ", " + airport_codes.iso_region.str.split('-', expand=True)[1]
airport_codes.loc[:, 'municipality'] = airport_codes.loc[:, 'municipality'] + ", " + airport_codes.iso_region.str.split('-', expand=True)[1]

airport_codes = airport_codes.drop_duplicates(subset=['municipality'])

# merge origin airport information 5 times
fw = fw.merge(airport_codes[['name', 'ident']].add_suffix('_origin_2'),
         how='left', left_on='origin', right_on='name_origin_2')

# merge similar columns
fw.loc[(fw.ICAO_code_origin.isna()), 'ident_origin_1'] = fw.loc[(fw.ICAO_code_origin.isna()), 'ident_origin_2'] 
fw.loc[(fw.ICAO_code_origin.isna()), 'ICAO_code_origin'] = fw.loc[(fw.ICAO_code_origin.isna()), 'ident_origin_2'] 

# Drop duplicated columns
fw = fw.drop(['ident_origin_2', 'name_origin_2'], axis=1)

# merge origin airport information 6 times
fw = fw.merge(airport_codes[['municipality', 'ident']].add_suffix('_origin_2'),
         how='left', left_on='origin', right_on='municipality_origin_2')

# merge similar columns
fw.loc[(fw.ICAO_code_origin.isna()), 'ident_origin_1'] = fw.loc[(fw.ICAO_code_origin.isna()), 'ident_origin_2'] 
fw.loc[(fw.ICAO_code_origin.isna()), 'ICAO_code_origin'] = fw.loc[(fw.ICAO_code_origin.isna()), 'ident_origin_2'] 

# Drop duplicated columns
fw = fw.drop(['ident_origin_2', 'municipality_origin_2'], axis=1)

# check for null values
fw.isna().sum()

#### Destination Location 
---

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# merge destination airport information 1 time
fw = fw.merge(airport_codes[['ident']].add_suffix('_destination_1'),
         how='left', left_on='ICAO_code_destination', right_on='ident_destination_1')

# Switch dual ICAO codes
fw.loc[(fw.ICAO_code_destination.notnull()) & (fw.ident_destination_1.isna()) & fw.destination.str.contains(' / '), 'ICAO_code_destination'] = fw.loc[(fw.ICAO_code_destination.notnull()) & (fw.ident_destination_1.isna()) & fw.destination.str.contains(' / '), 'destination'].str.split(' / ', n=1, expand=True)[1].str.replace(')', '')

# merge destination airport information 2 time
fw = fw.merge(airport_codes[['ident']].add_suffix('_destination_2'),
         how='left', left_on='ICAO_code_destination', right_on='ident_destination_2')

# merge similar columns
fw.loc[(fw.ICAO_code_destination.notnull()) & (fw.ident_destination_1.isna()) & (fw.destination.str.contains(' / ')), 'ident_destination_1'] = fw.loc[(fw.ICAO_code_destination.notnull()) & (fw.ident_destination_1.isna()) & (fw.destination.str.contains(' / ')), 'ident_destination_2']

# Drop duplicated column
fw = fw.drop(['ident_destination_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# drop duplicates in gps_code
airport_codes = airport_codes.drop_duplicates(subset=['gps_code'])

# merge destination airport information 3 time
fw = fw.merge(airport_codes[['gps_code', 'ident']].add_suffix('_destination_2'),
         how='left', left_on='ICAO_code_destination', right_on='gps_code_destination_2')

# merge similar columns
fw.loc[(fw.ICAO_code_destination.notnull()) & (fw.ident_destination_1.isna()), 'ident_destination_1'] = fw.loc[(fw.ICAO_code_destination.notnull()) & (fw.ident_destination_1.isna()), 'ident_destination_2']

# Drop duplicated columns
fw = fw.drop(['ident_destination_2', 'gps_code_destination_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# drop duplicates in local_code
airport_codes = airport_codes.drop_duplicates(subset=['local_code'])

# merge destination airport information 4 times
fw = fw.merge(airport_codes[['local_code', 'ident']].add_suffix('_destination_2'),
         how='left', left_on='ICAO_code_destination', right_on='local_code_destination_2')

# merge similar columns
fw.loc[(fw.ICAO_code_destination.notnull()) & (fw.ident_destination_1.isna()), 'ident_destination_1'] = fw.loc[(fw.ICAO_code_destination.notnull()) & (fw.ident_destination_1.isna()), 'ident_destination_2']

# Drop duplicated columns
fw = fw.drop(['ident_destination_2', 'local_code_destination_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# split coordinates and convert to numeric
airport_codes["Latitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[1])
airport_codes["Longitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[0])

airport_codes = airport_codes.replace(regex={r' Heliport': '', r' Airpark':'', r' Airport': '', r' Regional':'', 
                                             r' Municipal':'', r' International':'', r' LRRS':''})

airport_codes.loc[:, 'name'] = airport_codes.loc[:, 'name'] + ", " + airport_codes.iso_region.str.split('-', expand=True)[1]
airport_codes.loc[:, 'municipality'] = airport_codes.loc[:, 'municipality'] + ", " + airport_codes.iso_region.str.split('-', expand=True)[1]

airport_codes = airport_codes.drop_duplicates(subset=['municipality'])

# merge destination airport information 5 times
fw = fw.merge(airport_codes[['name', 'ident']].add_suffix('_destination_2'),
         how='left', left_on='destination', right_on='name_destination_2')

# merge similar columns
fw.loc[(fw.ICAO_code_destination.isna()), 'ident_destination_1'] = fw.loc[(fw.ICAO_code_destination.isna()), 'ident_destination_2'] 
fw.loc[(fw.ICAO_code_destination.isna()), 'ICAO_code_destination'] = fw.loc[(fw.ICAO_code_destination.isna()), 'ident_destination_2'] 

# Drop duplicated columns
fw = fw.drop(['ident_destination_2', 'name_destination_2'], axis=1)

# merge destination airport information 6 times
fw = fw.merge(airport_codes[['municipality', 'ident']].add_suffix('_destination_2'),
         how='left', left_on='destination', right_on='municipality_destination_2')

# merge similar columns
fw.loc[(fw.ICAO_code_destination.isna()), 'ident_destination_1'] = fw.loc[(fw.ICAO_code_destination.isna()), 'ident_destination_2'] 
fw.loc[(fw.ICAO_code_destination.isna()), 'ICAO_code_destination'] = fw.loc[(fw.ICAO_code_destination.isna()), 'ident_destination_2'] 

# Drop duplicated columns
fw = fw.drop(['ident_destination_2', 'municipality_destination_2'], axis=1)

# check for null values
fw.isna().sum()

#### Coordinates merging
---

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# split coordinates and convert to numeric
airport_codes["Latitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[1])
airport_codes["Longitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[0])

# merge by origin
fw = fw.merge(airport_codes.add_suffix('_origin'),
         how='left', left_on='ident_origin_1', right_on='ident_origin')

# replace origin_Latitude and origin_Longitude
fw.loc[fw.origin_Latitude.isna(), 'origin_Latitude'] = fw.loc[fw.origin_Latitude.isna(), 'Latitude_origin']
fw.loc[fw.origin_Longitude.isna(), 'origin_Longitude'] = fw.loc[fw.origin_Longitude.isna(), 'Longitude_origin']

# drop all useless columns
fw = fw.drop(['coordinates_origin', 'Latitude_origin', 'Longitude_origin', 'continent_origin', 'elevation_ft_origin', 'type_origin',
         'local_code_origin', 'gps_code_origin', 'municipality_origin', 'iso_region_origin', 'ident_origin', 'iata_code_origin'], axis=1)

# merge by destination
fw = fw.merge(airport_codes.add_suffix('_destination'),
         how='left', left_on='ident_destination_1', right_on='ident_destination')

# replace destination_Latitude and destination_Longitude
fw.loc[fw.destination_Latitude.isna(), 'destination_Latitude'] = fw.loc[fw.destination_Latitude.isna(), 'Latitude_destination']
fw.loc[fw.destination_Longitude.isna(), 'destination_Longitude'] = fw.loc[fw.destination_Longitude.isna(), 'Longitude_destination']

# drop all useless columns
fw = fw.drop(['coordinates_destination', 'Latitude_destination', 'Longitude_destination', 'continent_destination', 'elevation_ft_destination', 'type_destination',
         'local_code_destination', 'gps_code_destination', 'municipality_destination', 'iso_region_destination', 'ident_destination', 'iata_code_destination'], axis=1)

# check for null values
fw.isna().sum()

#### Additional Information
---

# add Owner column
fw = fw.merge(fw_info[['Tail_number','Owner']], how='left', left_on='tail_number', right_on='Tail_number').drop('Tail_number', axis=1)

# calculate the distance
for i in range(len(fw)):
    try:
        p1 = Point(str(fw.origin_Latitude[i]) + ' ' + str(fw.origin_Longitude[i]))
        p2 = Point(str(fw.destination_Latitude[i]) + ' ' + str(fw.destination_Longitude[i]))
    
        fw.loc[i, 'distance_mi'] = distance.distance(p1,p2).miles
    except:
        pass

# split states 
fw['origin_state'] = fw.origin.str.split(', ', n=1, expand=True)[1]
fw.loc[fw.origin_state.isna(), 'origin_state'] = fw.origin_location.str.split(', ', n=1, expand=True)[1].str.rsplit(') ', n=1, expand=True)[0]

fw['destination_state'] = fw.destination.str.split(', ', n=1, expand=True)[1]
fw.loc[fw.destination_state.isna(), 'destination_state'] = fw.destination_location.str.split(', ', n=1, expand=True)[1].str.rsplit(') ', n=1, expand=True)[0]

# split cities
fw['city_origin'] = fw.origin_location.str.rsplit(" (", n=1, expand=True)[1].str.split(")", n=1, expand=True)[0].str.split(',', n=1, expand=True)[0]
fw.loc[fw.city_origin.isna(), 'city_origin'] =fw.loc[fw.city_origin.isna(), 'origin'].str.split(',', n=1, expand=True)[0]

fw['city_destination'] = fw.destination_location.str.rsplit(" (", n=1, expand=True)[1].str.split(")", n=1, expand=True)[0].str.split(',', n=1, expand=True)[0]
fw.loc[fw.city_destination.isna(), 'city_destination'] = fw.loc[fw.city_destination.isna(), 'destination'].str.split(',', n=1, expand=True)[0]

# add shift
fw['departure_shift'] = np.where((pd.to_datetime(fw.departure.str.split(' ', n=1, expand=True)[0]).dt.time.astype('str') < '19:00:00') &\
                       (pd.to_datetime(fw.departure.str.split(' ', n=1, expand=True)[0]).dt.time.astype('str') > '07:00:00'), 'day', 'night')

fw['arrival_shift'] = np.where((pd.to_datetime(fw.arrival.str.split(' ', n=1, expand=True)[0]).dt.time.astype('str') < '19:00:00') &\
                       (pd.to_datetime(fw.arrival.str.split(' ', n=1, expand=True)[0]).dt.time.astype('str') > '07:00:00'), 'day', 'night')

# rename columns and drop unnecessary columns
fw = fw.rename({'ident_origin_1': 'origin_code', 'ident_destination_1': 'destination_code', 
           'name_origin':'airport_origin', 'iso_country_origin': 'country_origin',
          'name_destination':'airport_destination', 'iso_country_destination': 'country_destination'}, axis=1)\
       .drop(['origin_location', 'destination_location', 'ICAO_code_origin', 'ICAO_code_destination'], axis=1)

fw.info()

# rearrange columns
fw = fw.iloc[:,:3].join([fw.iloc[:,12], fw.iloc[:,3], fw.iloc[:,13], fw.iloc[:,4], fw.iloc[:,5:8], fw.iloc[:,19], fw.iloc[:,18], 
                    fw.iloc[:,8:12], fw.iloc[:,14], fw.iloc[:,22], fw.iloc[:,20], fw.iloc[:,15],
                    fw.iloc[:,16], fw.iloc[:,23], fw.iloc[:,21], fw.iloc[:,17], fw.iloc[:,24:26]])

fw.head()

fw.shape

# check for null values
fw.isna().sum()

# number of existing tail numbers with flight history
fw.tail_number.nunique() 

# number of existing tail numbers with flight history
fw.Owner.nunique() 

# # import to csv
# fw.to_csv("..//Datasets/NEW/FW_with_airports.csv", index=False)
