==> FW/Scripts/2. FW METAR data merging.ipynb <==
import os
import sys
import pandas as pd
import numpy as np
sys.path.append(os.getcwd() + '/Modules/')
from utc_converter import utc_converter
from metar_decoder import metar_decoder
pd.set_option('display.max_columns', None)

#### Data preprocessing
---

# load data frame
fw = pd.read_csv("..//Datasets/NEW/FW_with_airports.csv")

# add time with UTC format 
fw = utc_converter(fw)

# convert date to datetime
fw.date = pd.to_datetime(fw.date, format='%Y-%m-%d')

# split ICAO_code
fw['ICAO_origin'] = fw.origin.str.rsplit("(", n = 1, expand = True)[1].str.split(")", n = 1, expand = True)[0]
fw['ICAO_destination'] = fw.destination.str.rsplit("(", n = 1, expand = True)[1].str.split(")", n = 1, expand = True)[0]

# Split dual ICAO and take the last one
fw.loc[fw.ICAO_origin.str.contains(' / ', na=False), 'ICAO_origin'] = fw.loc[fw.ICAO_origin.str.contains(' / ', na=False), 'ICAO_origin']\
                                                                        .str.split(" / ", expand=True, n=1)[1]

fw.loc[fw.ICAO_destination.str.contains(' / ', na=False), 'ICAO_destination'] = fw.loc[fw.ICAO_destination.str.contains(' / ', na=False), 'ICAO_destination']\
                                                                                  .str.split(" / ", expand=True, n=1)[1]

# replace null ICAO code by airport code
fw.loc[fw.ICAO_origin.isna(), 'ICAO_origin'] = fw.loc[fw.ICAO_origin.isna(), 'origin_code'] 
fw.loc[fw.ICAO_destination.isna(), 'ICAO_destination'] = fw.loc[fw.ICAO_destination.isna(), 'destination_code']

# Transform aita code to icao
fw.loc[~fw.ICAO_origin.str.startswith('K', na=False), 'ICAO_origin'] = 'K' + fw.loc[~fw.ICAO_origin.str.startswith('K', na=False), 'ICAO_origin']
fw.loc[~fw.ICAO_destination.str.startswith('K', na=False), 'ICAO_destination'] = 'K' + fw.loc[~fw.ICAO_destination.str.startswith('K', na=False), 'ICAO_destination']

# extract hour from dep_UTC_time and assign a new variable
fw['dep_UTC_hour'] = pd.to_datetime(fw.dep_UTC_time.astype('str'), format='%H:%M:%S', errors='coerce').dt.hour.astype('str')
fw['arr_UTC_hour'] = pd.to_datetime(fw.arr_UTC_time.astype('str'), format='%H:%M:%S', errors='coerce').dt.hour.astype('str')

# sort by 'tail_number', 'date', 'dep_UTC_hour'
fw = fw.sort_values(['tail_number', 'date', 'dep_UTC_hour']).reset_index(drop=True)

# inspect data
fw

#### METAR data merging for Origin Location
---

# scrape METAR data for Origin
from metar_scraper_origin import metar_scraper_origin
metar_scraper_origin(fw)

# load data frame
metar_origin = pd.read_csv("..//Datasets/NEW/METAR/origin/FW_METAR_origin.csv")

# split date and time. Extract only hours  
metar_origin['hour'] = pd.to_datetime(metar_origin.date, format='%d/%m/%Y %H:%M').dt.hour.astype('str') + '.0'
metar_origin['date'] = pd.to_datetime(metar_origin.date, format='%d/%m/%Y %H:%M').dt.date.astype('str')

# convert date to string
fw.date = fw.date.astype('str')

# sort by 'tail_number', 'date', 'hour' and drop duplicates
metar_origin = metar_origin.sort_values(['tail_number', 'date', 'hour']).drop_duplicates(subset=['tail_number', 'ICAO_origin', 'date', 'hour'])\
                           .reset_index(drop=True)

# inspect data
metar_origin.head()

# merge by 'tail_number', 'date', 'ICAO_origin', 'dep_UTC_hour'
fw = fw.merge(metar_origin.add_suffix('_origin'), how='left', left_on=['tail_number', 'date', 'ICAO_origin', 'dep_UTC_hour'],
        right_on=['tail_number_origin', 'date_origin', 'ICAO_origin_origin', 'hour_origin'])\
        .drop(['ICAO_origin', 'dep_UTC_hour','tail_number_origin', 'ICAO_origin_origin', 'date_origin', 'hour_origin'], axis=1)

# inspect merging
fw.head()

# check for null values
fw.METAR_origin.isna().sum()

fw.loc[fw.METAR_origin.isna()]

#### METAR data merging for Destination Location
---

# scrape METAR data for Origin
from metar_scraper_destination import metar_scraper_destination
metar_scraper_destination(fw)

# load data frame
metar_destination = pd.read_csv("../Datasets/METAR/FW/destination/FW_METAR_destination.csv")

# split date and time. Extract only hours  
metar_destination['hour'] = pd.to_datetime(metar_destination.date, format='%d/%m/%Y %H:%M').dt.hour.astype('str') + '.0'
metar_destination['date'] = pd.to_datetime(metar_destination.date, format='%d/%m/%Y %H:%M').dt.date.astype('str')

# sort by 'tail_number', 'date', 'hour' and drop duplicates
metar_destination = metar_destination.sort_values(['tail_number', 'date', 'hour']).drop_duplicates(subset=['tail_number', 'ICAO_destination', 'date', 'hour'])\
                           .reset_index(drop=True)

# inspect data
metar_destination.head()

# merge by 'tail_number', 'date', 'ICAO_origin', 'arr_UTC_hour'
fw = fw.merge(metar_destination.add_suffix('_destination'), how='left', left_on=['tail_number', 'date', 'ICAO_destination', 'arr_UTC_hour'],
        right_on=['tail_number_destination', 'date_destination', 'ICAO_destination_destination', 'hour_destination'])\
        .drop(['ICAO_destination', 'arr_UTC_hour','tail_number_destination', 'ICAO_destination_destination', 
               'date_destination', 'hour_destination'], axis=1)

# inspect data
fw.head()

# check for null values
fw.METAR_destination.isna().sum()

#### METAR Decoding
---

# decode metar data
fw = metar_decoder(fw)

fw = fw.loc[:,:'METAR_origin'].join([fw.loc[:,'METAR_time_origin':'remarks_origin'], 
                                fw.loc[:,'METAR_destination'], fw.loc[:,'METAR_time_destination':'remarks_destination']])

# inspect data
fw.head()

fw.info()

fw.isna().sum()

# export as csv
fw.to_csv('../Datasets/NEW/fw_with_metar.csv', index=False)
