==> FW/Scripts/3 FW_anomaly.ipynb <==
#load packages
import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt
pd.set_option('display.max_columns', 999)

fw = pd.read_csv("../Datasets/NEW/FW_with_airports.csv", parse_dates=['date'])

fw = fw.iloc[:,:11].join(fw.iloc[:,-1])

fw.head()

fw.info()

fw.isna().sum()

# exclude all diverted and practice flights 
fw = fw.loc[(fw.origin != fw.destination) & (fw.arrival.notnull()) & (fw.destination.notnull()) & (fw.duration.notnull())
           & (fw.distance_mi.notnull())]

# convert duration to minutes and add a new column 
fw['duration_minutes'] = pd.to_datetime(fw.duration, format="%H:%M", errors='coerce').dt.hour * 60 + pd.to_datetime(fw.duration, format="%H:%M",errors='coerce').dt.minute

# exclude all flight less that 2 minutes. We cansider that as diverted or canceled 
fw = fw.loc[fw.duration_minutes > 2]

# set Owner and tail_number as indices
fw = fw.set_index(['Owner', 'tail_number'])

# calculate 25 quantile of each tail number of the airmedical service
fw['duration_q_25'] = fw.groupby(['Owner', 'tail_number']).duration_minutes.quantile(0.25)

# calculate 75 quantile of each tail number of the airmedical service
fw['duration_q_75'] = fw.groupby(['Owner', 'tail_number']).duration_minutes.quantile(0.75)

# calculate IQR
fw['duration_iqr'] = fw['duration_q_75'] - fw['duration_q_25']

# calculate lower whisker
fw['duration_lower_whisker'] = fw['duration_q_25'] - (1.5 * fw['duration_iqr'])

# calculate upper whisker
fw['duration_upper_whisker'] = fw['duration_q_75'] + (1.5 * fw['duration_iqr'])

fw['duration_outlier'] = np.where((fw.duration_minutes < fw.duration_lower_whisker) | (fw.duration_minutes > fw.duration_upper_whisker), "Yes", "No")

# calculate 25 quantile of each tail number of the airmedical service
fw['distance_q_25'] = fw.groupby(['Owner', 'tail_number']).distance_mi.quantile(0.25)

# calculate 75 quantile of each tail number of the airmedical service
fw['distance_q_75'] = fw.groupby(['Owner', 'tail_number']).distance_mi.quantile(0.75)

# calculate IQR
fw['distance_iqr'] = fw['distance_q_75'] - fw['distance_q_25']

# calculate lower whisker
fw['distance_lower_whisker'] = fw['distance_q_25'] - (1.5 * fw['distance_iqr'])

# calculate upper whisker
fw['distance_upper_whisker'] = fw['distance_q_75'] + (1.5 * fw['distance_iqr'])

fw['distance_outlier'] = np.where((fw.distance_mi < fw.distance_lower_whisker) | (fw.distance_mi > fw.distance_upper_whisker), "Yes", "No")

fw.head()

fw.to_csv('..//Datasets/NEW/FW_Outliers.csv')
