==> FW/Scripts/5. FW Tracker.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', None)

# import data
fw = pd.read_csv("..//Datasets/NEW/FW_with_airports.csv")
fw_info = pd.read_csv("..//Datasets/NEW/FW_services_info.csv")

fw.head()

# number of existing tail numbers with flight history
fw.tail_number.nunique() 

# The percentage of the total tail numbers
print(round(fw.tail_number.nunique() / fw_info.Tail_number.nunique() * 100, 2), "%")

fw.info()

fw = fw.iloc[::-1].reset_index(drop=True)

tail_number = []
date = []
time = []
airport = []
location = []
duration = []
distance_mi = []
Latitude = []
Longitude= []
Owner = []

for i in range(len(fw)):
    if fw.tail_number[i] == fw.tail_number.iloc[i-1]: 
        tail_number += [fw.tail_number[i]]
        date += [fw.date[i] + ' ' + str(fw.departure[i])]
        airport += [fw.airport_origin[i]]
        location += [fw.city_origin[i]]
        Latitude += [fw.origin_Latitude[i]]
        Longitude += [fw.origin_Longitude[i]]
        Owner += [fw.Owner[i]]
        duration += [fw.duration[i-1]]
        distance_mi += [fw.distance_mi[i-1]]
    else:
        tail_number += [fw.tail_number[i]]
        date += [fw.date[i] + ' ' + str(fw.departure[i])]
        airport += [fw.airport_origin[i]]
        location += [fw.city_origin[i]]
        Latitude += [fw.origin_Latitude[i]]
        Longitude += [fw.origin_Longitude[i]]
        Owner += [fw.Owner[i]]
        duration += [0]
        distance_mi += [0]
        
#     time += ['Departure']
    
#     tail_number += [fw.tail_number[i]]
#     date += [fw.date[i] + ' ' + str(fw.arrival[i])]
#     airport += [fw.airport_destination[i]]
#     location += [fw.city_destination[i]]
#     distance_mi += [fw.distance_mi[i]]
#     Latitude += [fw.destination_Latitude[i]]
#     Longitude += [fw.destination_Longitude[i]]
#     Owner += [fw.Owner[i]]
#     time += ['Arrival']
    
fw_tracker = pd.DataFrame(list(zip(tail_number, date, airport, location, duration, distance_mi, Latitude, Longitude, Owner)), 
             columns= ['tail_number', 'date', 'airport', 'location', 'duration', 'distance_mi', 'Latitude', 'Longitude', 'Owner'])    

# drop inconsequences
index = []
for i in range(len(fw_tracker)-1):
    if fw_tracker.airport.iloc[i] == fw_tracker.airport.iloc[i+1]:
        index += [i]
        
fw_tracker = fw_tracker.drop(index).reset_index(drop=True)

fw_tracker

fw_tracker.to_csv('..//Datasets/NEW/fw_tracker.csv', index=False)

# create a dict
tail_number = {}

# add home location and total flights
for i in fw_tracker.tail_number.unique():
    tail_number[i] = [fw_tracker.loc[fw_tracker.tail_number == i].airport.value_counts().index[0]]  
    tail_number[i].append(len(fw_tracker.loc[fw_tracker.tail_number == i]))

# create containers    
one_loc = 0
two_loc = 0
three_loc = 0
four_loc = 0

# loop over all paths 
for i in range(len(fw_tracker)):
    if fw_tracker.tail_number[i] == fw_tracker.tail_number.iloc[i-1]:
        if fw_tracker.airport[i] == fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and fw_tracker.airport.iloc[i-2] == fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]:
            one_loc += 1
        
        if fw_tracker.airport[i] == fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and fw_tracker.airport.iloc[i-3] == fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]:
            two_loc += 1
            
        if fw_tracker.airport[i] == fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and fw_tracker.airport.iloc[i-4] == fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and fw_tracker.airport.iloc[i-2] != fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]:
            three_loc += 1
            
        if fw_tracker.airport[i] == fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and fw_tracker.airport.iloc[i-5] == fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and fw_tracker.airport.iloc[i-2] != fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and fw_tracker.airport.iloc[i-3] != fw_tracker.loc[fw_tracker.tail_number == fw_tracker.tail_number[i]].airport.value_counts().index[0]:
            four_loc += 1
    else:
        tail_number[fw_tracker.tail_number.iloc[i-1]].append(one_loc)
        tail_number[fw_tracker.tail_number.iloc[i-1]].append(two_loc)
        tail_number[fw_tracker.tail_number.iloc[i-1]].append(three_loc)
        tail_number[fw_tracker.tail_number.iloc[i-1]].append(four_loc)
        
        # reset containers
        one_loc = 0
        two_loc = 0
        three_loc = 0
        four_loc = 0

fw_hops = pd.DataFrame.from_dict(tail_number, orient='index',
                       columns=['home_loc', 'total_flights','one_loc', 'two_loc', 'three_loc', 'four_loc'])\
            .reset_index().rename({'index': 'tail_number'}, axis=1)

fw_hops = fw_hops.merge(fw.drop_duplicates(subset=['city_origin', 'origin_state'])[['airport_origin', 'city_origin', 'origin_state']],
             how='left', left_on='home_loc', right_on='airport_origin').drop('airport_origin', axis=1)

# add owner
fw_hops = fw_hops.merge(fw_info[['Tail_number', 'Owner']], how='left',
             left_on='tail_number', right_on='Tail_number').drop('Tail_number', axis=1)

fw_hops

fw_hops.drop([0, 176], axis=0).to_csv('..//Datasets/NEW/fw_hops.csv', index=False)
