==> FW/Scripts/6.1 FW_IHC.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 999)

# load data frame
fw = pd.read_csv("..//Datasets/NEW/FW_with_airports.csv")

# number of the unique air medical services
fw.Owner.nunique()

# take only IHC service
ihc = fw.loc[fw.Owner == 'IHC HEALTH SERVICES INCSALT LAKE CITY, UT, US(Corporation)'].reset_index(drop=True)

ihc['duration_min'] = ihc.duration.str.split(":", expand=True)[0].astype('int') * 60 + ihc.duration.str.split(":", expand=True)[1].astype('int')

ihc['day_name'] = pd.to_datetime(ihc.date, format='%Y-%m-%d').dt.day_name()

ihc['weekends'] = np.where((ihc.day_name == 'Saturday') | (ihc.day_name == 'Sunday'), 1, 0)

# number of flights by tail number
ihc.tail_number.value_counts()

# average distance and duration by tail number
ihc.groupby('tail_number')[['distance_mi', 'duration_min']].mean()

# number of flights in origin and destination locations
ihc.groupby('tail_number')[['origin', 'destination']].nunique()

# top 5 of origin locations and number of flights for N381HC
ihc.groupby('tail_number').origin.value_counts().loc['N381HC'].head(5)

# top 5 of destination locations and number of flights for N381HC
ihc.groupby('tail_number').destination.value_counts().loc['N381HC'].head(5)

# top 5 of origin locations and number of flights for N481HC
ihc.groupby('tail_number').origin.value_counts().loc['N481HC'].head(5)

# top 5 of destination locations and number of flights for N481HC
ihc.groupby('tail_number').destination.value_counts().loc['N481HC'].head(5)

# distance summary statistics by departure shift
ihc.groupby('departure_shift').distance_mi.agg(['count','min','mean','median','max','std'])

# distance summary statistics by arrival shift
ihc.groupby('arrival_shift').distance_mi.agg(['count','min','mean','median','max','std'])

# distance summary statistics by day
ihc.groupby('day_name').distance_mi.agg(['count','min','mean','median','max','std'])

# duration summary statistics by departure shift
ihc.groupby('departure_shift').duration_min.agg(['count','min','mean','median','max','std'])

# duration summary statistics by arrival shift
ihc.groupby('arrival_shift').duration_min.agg(['count','min','mean','median','max','std'])

# duration summary statistics by day
ihc.groupby('day_name').duration_min.agg(['count','min','mean','median','max','std'])

# Salt Lake City flights
slc = fw.loc[(fw.origin == 'Salt Lake City Intl (KSLC)') | (fw.destination == 'Salt Lake City Intl (KSLC)')]

# competitors for N381HC in Salt Lake City as origin
pd.DataFrame(slc.loc[(slc.origin == 'Salt Lake City Intl (KSLC)') & (slc.tail_number != 'N481HC')].groupby('Owner')\
             .origin.value_counts().sort_values(ascending=False))

# competitors for N381HC in Salt Lake City as destination
pd.DataFrame(slc.loc[(slc.destination == 'Salt Lake City Intl (KSLC)') & (slc.tail_number != 'N481HC')].groupby('Owner')\
             .destination.value_counts().sort_values(ascending=False))

# competitors for N481HC in Salt Lake City as origin
pd.DataFrame(slc.loc[(slc.origin == 'Salt Lake City Intl (KSLC)') & (slc.tail_number != 'N381HC')].groupby('Owner')\
             .origin.value_counts().sort_values(ascending=False))

# competitors for N481HC in Salt Lake City as destination
pd.DataFrame(slc.loc[(slc.destination == 'Salt Lake City Intl (KSLC)') & (slc.tail_number != 'N381HC')].groupby('Owner')\
             .destination.value_counts().sort_values(ascending=False))

ihc.groupby('tail_number').destination.nunique()

# ihc.to_csv('ihc.csv', index=False)

# St George Rgnl (KSGU)
sgr = fw.loc[(fw.origin == 'St George Rgnl (KSGU)') | (fw.destination == 'St George Rgnl (KSGU)')]

# competitors for N381HC in St George Rgnl (KSGU) as origin
pd.DataFrame(sgr.loc[(sgr.origin == 'St George Rgnl (KSGU)') & (sgr.tail_number != 'N481HC')].groupby('Owner')\
             .origin.value_counts().sort_values(ascending=False))

# competitors for N381HC in St George Rgnl (KSGU) as destination
pd.DataFrame(sgr.loc[(sgr.destination == 'St George Rgnl (KSGU)') & (sgr.tail_number != 'N481HC')].groupby('Owner')\
             .destination.value_counts().sort_values(ascending=False))

# competitors for N481HC in St George Rgnl (KSGU) as origin
pd.DataFrame(sgr.loc[(sgr.origin == 'St George Rgnl (KSGU)') & (sgr.tail_number != 'N381HC')].groupby('Owner')\
             .origin.value_counts().sort_values(ascending=False))

# competitors for N481HC in St George Rgnl (KSGU) as destination
pd.DataFrame(sgr.loc[(sgr.destination == 'St George Rgnl (KSGU)') & (sgr.tail_number != 'N381HC')].groupby('Owner')\
             .destination.value_counts().sort_values(ascending=False))
