==> FW/Scripts/7.1 ILS and LPVorPV.ipynb <==
# Question: 
#### Is there correlation between airports used, Visibility/Ceiling, and ILS or LPV/LP?

1. Lateral Precision Performance with Vertical Guidance (LPV) is defined as an Approach with Vertical Guidance (APV); that is, an instrument approach based on a navigation system that is not required to meet the precision approach standards of ICAO Annex 10 but that provides both course and glidepath deviation information.                                                    

2. Localizer Performance without Vertical Guidance (LP) minima are added in locations where terrain or obstructions do not allow publication of vertically guided LPV minima. Lateral sensitivity increases as an aircraft get closer to the runway (or point in space for helicopters). LP is not a fail-down mode for LPV; LP and LPV are independent. LNAV is not a fail-down mode for LP. LP will not be published with lines of minima that contain approved vertical guidance (i.e. LNAV/VNAV or LPV). 

# import packages
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', None)

# import flights data with metar
fw = pd.read_csv('..//Datasets/NEW/FW_with_metar.csv')

# observe the dataset
fw.head()

fw.info()

# create a subset of data with visibility
fw_sub = fw[['origin_code', 'origin', 'destination_code', 'destination', 'visibility_origin', 'visibility_destination']]

# drop records
fw_sub = fw_sub.loc[fw_sub.origin_code != fw_sub.destination_code]

fw_sub.info()

# check for null values
fw_sub.isna().sum()

# drop null values and reset index
fw_sub = fw_sub.dropna().reset_index(drop=True)

fw_sub.shape

# restracture data
airport_code = []
location = []
visibility = []

for i in range(len(fw_sub)):
    airport_code += [fw_sub.origin_code.iloc[i]] 
    location += [fw_sub.origin.iloc[i]] 
    visibility += [fw_sub.visibility_origin.iloc[i]]
    
    airport_code += [fw_sub.destination_code.iloc[i]] 
    location += [fw_sub.destination.iloc[i]] 
    visibility += [fw_sub.visibility_destination.iloc[i]]

fw = pd.DataFrame(list(zip(airport_code, location, visibility)),
                 columns=['airport_code', 'location', 'visibility'])

# drop the duplicates
fw = fw.drop_duplicates()

# delete 'miles' word from visibility column
fw.loc[:,'visibility'] = fw.visibility.str.rsplit(' ', n=1, expand=True)[0]

# find out the non decimal numbers in visibility
fw[fw.visibility.str.contains('/', na=False)].visibility.unique()

# replace them to decimal in visibility
fw = fw.replace(regex={'missing':np.nan, 'less than ':''})
fw = fw.replace(regex={'2 1/2': '2.5', 
                      '1 1/2': '1.5', '1 3/4': '1.75', '1 1/4': '1.25',
                      '1/4': '0.25', '1/2': '0.5', '3/4': '0.75', 
                      '1/8': '0.125', '3/8':'0.375', '1/16':'0.0625'})

# convert visibility to numeric
fw.visibility = pd.to_numeric(fw.visibility)

# drop null values and reset index
fw = fw.dropna().reset_index(drop=True)

fw.head()

fw = fw.groupby(['airport_code', 'location']).agg(['count','mean', 'median', 'max', 'min', 'std'])\
       .rename({'count':'viz_count','mean':'avg_visibility', 'median':'median_visibility', 'max':'max_visibility',
               'min':'min_visibility', 'std': 'std_visibility'}, axis=1)

# drop multiindex columns name
fw.columns = fw.columns.droplevel()

# reset indexes
fw = fw.reset_index()

fw.head()

# import airport code data
airport_code =  pd.read_csv('..//Datasets/Airports/airport-codes.csv')

# check for duplicates in ident column
airport_code.ident.duplicated().sum()

# merge local code 
fw = fw.merge(airport_code[['ident','local_code']], how='left', left_on='airport_code', right_on='ident').drop('ident', axis=1)

fw.head()

# load airports infromation
airports = pd.read_csv('..//Datasets/Airports/FAA_Airport_Data.csv')

airports.info()

airports.head()

# observe the dataset
airports.isna().sum()

# drop columns which has a lot of null values
airports = airports.drop(['GPA','TCH', 'ILS Arpt Cntr (Y/N)', 'ILS CAT II or III'], axis=1)

# drop duplicates in FAA ID column
airports = airports.drop_duplicates(subset='FAA ID')

airports.head()

airports[airports.VIS.str.contains('/', na=False)].VIS.unique()

airports = airports.replace(regex={'2 1/2': '2.5', '1 1/8': '1.125', ' 7/8': '0.875', '1 3/8': '1.375', '2 3/4':'2.75',
                      '1 1/2': '1.5', '1 3/4': '1.75', '1 1/4': '1.25', '1 5/8': '1.625', ' 3/5':'0.6', '2 1/4': '2.25',
                      '1/4': '0.25', '1/2': '0.5', '3/4': '0.75', '1 7/8': '1.875', 
                      '1/8': '0.125', '3/8':'0.375', '1/16':'0.0625',' 5/8':'0.625', ' 2/3': '0.67'})

airports.shape

airports['FAA ID'].isin(fw.local_code).sum()

# merge by local code
fw = fw.merge(airports, how='inner', left_on='local_code', right_on='FAA ID')

fw.info()

fw.isna().sum()

# sort by count to observe more diverse airports
fw = fw.sort_values('viz_count' , ascending=False)

# reset indexes
fw = fw.reset_index(drop=True)

# fw = fw.dropna(subset=fw.columns[9:])

fw.shape

fw.head()

fw.drop(['local_code','FAA ID', 'location'], axis=1).to_csv('../Datasets/NEW/FW_airport_analysis.csv', index=False)

ils_air = fw[['ILS @ Arpt (Y/N)', 'avg_visibility']].replace(regex={"Y":1, "N":0})

ils_air = ils_air.loc[ils_air.avg_visibility < 10]

from sklearn import tree
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from sklearn.metrics import precision_score, recall_score

# Split X and Y into training and testing datasets
train_X, test_X, train_Y, test_Y = train_test_split(ils_air['avg_visibility'], ils_air['ILS @ Arpt (Y/N)'], test_size=0.25, random_state=42)

# Ensure training dataset has only 75% of original X data
print(train_X.shape[0] / ils_air['avg_visibility'].shape[0])

# Ensure testing dataset has only 25% of original X data
print(test_Y.shape[0] / ils_air['ILS @ Arpt (Y/N)'].shape[0])

# Initialize decision tree classifier
mytree = tree.DecisionTreeClassifier()

# Fit the decision tree on training data
mytree.fit(pd.DataFrame(train_X), train_Y)

# Predict churn labels on testing data
pred_test_Y = mytree.predict(pd.DataFrame(test_X))

# Calculate accuracy score on testing data
test_accuracy = accuracy_score(test_Y, pred_test_Y)

# Print test accuracy
print('Test accuracy:', round(test_accuracy, 4))

depth_list = [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]

depth_tuning = np.array([[ 2.,  0.],
       [ 3.,  0.],
       [ 4.,  0.],
       [ 5.,  0.],
       [ 6.,  0.],
       [ 7.,  0.],
       [ 8.,  0.],
       [ 9.,  0.],
       [10.,  0.],
       [11.,  0.],
       [12.,  0.],
       [13.,  0.],
       [14.,  0.]])

# Run a for loop over the range of depth list length
for index in range(0, len(depth_list)):
  # Initialize and fit decision tree with the `max_depth` candidate
  mytree = tree.DecisionTreeClassifier(max_depth=depth_list[index])
  mytree.fit(pd.DataFrame(train_X), train_Y)
  # Predict churn on the testing data
  pred_test_Y = mytree.predict(pd.DataFrame(test_X))
  # Calculate the recall score 
  depth_tuning[index,1] = accuracy_score(test_Y, pred_test_Y)

# Name the columns and print the array as pandas DataFrame
col_names = ['Max_Depth','accuracy_score']
print(pd.DataFrame(depth_tuning, columns=col_names))

# Initialize decision tree classifier
mytree = tree.DecisionTreeClassifier(max_depth=3)

# Fit the decision tree on training data
mytree.fit(pd.DataFrame(train_X), train_Y)

# Predict churn labels on testing data
pred_test_Y = mytree.predict(pd.DataFrame(test_X))

# Calculate accuracy score on testing data
test_accuracy = accuracy_score(test_Y, pred_test_Y)

# Print test accuracy
print('Test accuracy:', round(test_accuracy, 4))
